Apply DTL_DEFAULT to core init. Move the header including common dtl definitions into common
