+++ src/dyad/dtl/dyad_dtl_api.c
@@ -4,7 +4,7 @@
 #error "no config"
 #endif
 
-#include <dyad/dtl/dyad_dtl_impl.h>
+#include <dyad/dtl/dyad_dtl_api.h>
 #include <dyad/dtl/flux_dtl.h>
 #include <dyad/common/dyad_logging.h>
 #include <dyad/common/dyad_profiler.h>
