+++ src/dyad/dtl/ucx_dtl.h
@@ -7,7 +7,7 @@
 #error "no config"
 #endif
 
-#include <dyad/dtl/dyad_dtl_impl.h>
+#include <dyad/dtl/dyad_dtl_api.h>
 #include <dyad/dtl/ucx_ep_cache.h>
 #include <ucp/api/ucp.h>
 #include <stdlib.h>
