+++ src/dyad/modules/dyad.c
@@ -16,7 +16,7 @@
 
 #include <dyad/common/dyad_envs.h>
 #include <dyad/common/dyad_rc.h>
-#include <dyad/dtl/dyad_dtl_impl.h>
+#include <dyad/dtl/dyad_dtl_api.h>
 #include <dyad/core/dyad_core.h>
 #include <dyad/common/dyad_logging.h>
 #include <dyad/common/dyad_profiler.h>
